add reset of test device
this exists because in 1.6 boards there is too much capacitance so that the device can actually stay 'alive' if left turned on. this means that in phase2/3 of testing the pulse it will not cause the jig to detect the device has been turned off/on if left plugged in for a while because it hasn't! so we add a button that allows the user to manually force a reset. this should be an unusual situation but we want to account for it...